Fix running kubernetes watcher unit tests
